feat: API 관련 타입 정의 추가
