--- src/app/lib/type.ts
@@ -1,7 +1,0 @@
-export interface RecommendationResponse {
-  title: string;
-  position: string;
-  category: string;
-  image: string;
-  url: string;
-}
