+++ src/lib/axiosInstance.ts
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const axiosInstance = axios.create({
+  baseURL: process.env.NEXT_PUBLIC_API_URL,
+});
+
+export default axiosInstance;
